fix old setDefaultDir flag part in add subcommand part, change minute type to float

+++ backup/parser.py
@@ -20,7 +20,7 @@ def create_parser() -> argparse.ArgumentParser:
 
   add_parser = subparsers.add_parser("add", help="add -h")
   add_parser.add_argument('target', nargs=1, type=Path)
-  add_parser.add_argument('interval', nargs=1, type=int)
+  add_parser.add_argument('interval', nargs=1, type=float)
   add_parser.add_argument('directory', nargs='?', type=Path, const=None)
 
   config_parser = subparsers.add_parser("config", help="config -h")
